backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==3.2.0",
    "email-validator>=2.3.0",
    "fastapi>=0.133.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
